SFA-pytorch/requirements.txt
argparse==1.4.0
h5py==2.10.0
Pillow==6.2.1
numpy==1.17.3
scipy==1.3.1
scikit-learn==0.19.1
torch==1.3.0
torchvision==0.4.1